LDE-714: moved tests for update_cube
